Add checkbox and api call for private marvel cdb decks
When loading a deck from marvel cdb, there is a separate api path fro getting private decks and public decks. This checkbox passes a boolean value down the stack to tell the function which path to use. For lotr, there is no private deck API, so we hide the checkbox by changing the game properties to see if there is a private api url configured.